+++ Firestore/Source/API/FIRFirestore.mm
@@ -19,7 +19,6 @@
 #import <FirebaseCore/FIRApp.h>
 #import <FirebaseCore/FIRAppInternal.h>
 #import <FirebaseCore/FIRComponentContainer.h>
-#import <FirebaseCore/FIROptions.h>
 
 #include <memory>
 #include <string>
@@ -29,7 +28,6 @@
 
 #import "Firestore/Source/API/FIRCollectionReference+Internal.h"
 #import "Firestore/Source/API/FIRDocumentReference+Internal.h"
-#import "Firestore/Source/API/FIRFirestore+Internal.h"
 #import "Firestore/Source/API/FIRListenerRegistration+Internal.h"
 #import "Firestore/Source/API/FIRQuery+Internal.h"
 #import "Firestore/Source/API/FIRTransaction+Internal.h"
@@ -65,8 +63,13 @@ using firebase::firestore::core::EventListener;
 using firebase::firestore::model::DatabaseId;
 using firebase::firestore::util::AsyncQueue;
 using firebase::firestore::util::Empty;
+using firebase::firestore::util::MakeCallback;
+using firebase::firestore::util::MakeNSError;
+using firebase::firestore::util::MakeNSString;
+using firebase::firestore::util::MakeString;
 using firebase::firestore::util::ObjcThrowHandler;
 using firebase::firestore::util::SetThrowHandler;
+using firebase::firestore::util::Status;
 using firebase::firestore::util::StatusOr;
 using firebase::firestore::util::ThrowIllegalState;
 using firebase::firestore::util::ThrowInvalidArgument;
@@ -102,11 +105,11 @@ NS_ASSUME_NONNULL_BEGIN
     ThrowIllegalState("Failed to get FirebaseApp instance. Please call FirebaseApp.configure() "
                       "before using Firestore");
   }
-  return [self firestoreForApp:app database:util::MakeNSString(DatabaseId::kDefault)];
+  return [self firestoreForApp:app database:MakeNSString(DatabaseId::kDefault)];
 }
 
 + (instancetype)firestoreForApp:(FIRApp *)app {
-  return [self firestoreForApp:app database:util::MakeNSString(DatabaseId::kDefault)];
+  return [self firestoreForApp:app database:MakeNSString(DatabaseId::kDefault)];
 }
 
 // TODO(b/62410906): make this public
@@ -142,7 +145,7 @@ NS_ASSUME_NONNULL_BEGIN
 
     FSTPreConverterBlock block = ^id _Nullable(id _Nullable input) {
       if ([input isKindOfClass:[FIRDocumentReference class]]) {
-        FIRDocumentReference *documentReference = (FIRDocumentReference *)input;
+        auto documentReference = (FIRDocumentReference *)input;
         return [[FSTDocumentKeyReference alloc] initWithKey:documentReference.key
                                                  databaseID:documentReference.firestore.databaseID];
       } else {
@@ -183,7 +186,7 @@ NS_ASSUME_NONNULL_BEGIN
   }
 
   return [[FIRCollectionReference alloc]
-      initWithReference:_firestore->GetCollection(util::MakeString(collectionPath))];
+      initWithReference:_firestore->GetCollection(MakeString(collectionPath))];
 }
 
 - (FIRDocumentReference *)documentWithPath:(NSString *)documentPath {
@@ -194,7 +197,7 @@ NS_ASSUME_NONNULL_BEGIN
     ThrowInvalidArgument("Invalid path (%s). Paths must not contain // in them.", documentPath);
   }
 
-  DocumentReference documentReference = _firestore->GetDocument(util::MakeString(documentPath));
+  DocumentReference documentReference = _firestore->GetDocument(MakeString(documentPath));
   return [[FIRDocumentReference alloc] initWithReference:std::move(documentReference)];
 }
 
@@ -207,7 +210,7 @@ NS_ASSUME_NONNULL_BEGIN
                          collectionID);
   }
 
-  auto query = _firestore->GetCollectionGroup(util::MakeString(collectionID));
+  auto query = _firestore->GetCollectionGroup(MakeString(collectionID));
   return [[FIRQuery alloc] initWithQuery:std::move(query) firestore:_firestore];
 }
 
@@ -253,16 +256,16 @@ NS_ASSUME_NONNULL_BEGIN
           if (error.domain != FIRFirestoreErrorDomain) {
             internalTransaction->MarkPermanentlyFailed();
           }
-          internalCallback(util::Status::FromNSError(error));
+          internalCallback(Status::FromNSError(error));
         } else {
-          internalCallback(util::Status::OK());
+          internalCallback(Status::OK());
         }
       });
     }
 
-    void HandleFinalStatus(const util::Status &status) {
+    void HandleFinalStatus(const Status &status) {
       if (!status.ok()) {
-        user_completion_(nil, util::MakeNSError(status));
+        user_completion_(nil, MakeNSError(status));
         return;
       }
 
@@ -293,7 +296,7 @@ NS_ASSUME_NONNULL_BEGIN
   //
   // PORTING NOTE: Other platforms where the user return value is internally representable don't
   // need this wrapper.
-  auto objcTranslator = [result_capture](const util::Status &status) {
+  auto objcTranslator = [result_capture](const Status &status) {
     result_capture->HandleFinalStatus(status);
   };
 
@@ -319,26 +322,26 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 - (void)enableNetworkWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
-  _firestore->EnableNetwork(util::MakeCallback(completion));
+  _firestore->EnableNetwork(MakeCallback(completion));
 }
 
 - (void)disableNetworkWithCompletion:(nullable void (^)(NSError *_Nullable))completion {
-  _firestore->DisableNetwork(util::MakeCallback(completion));
+  _firestore->DisableNetwork(MakeCallback(completion));
 }
 
 - (void)clearPersistenceWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
-  _firestore->ClearPersistence(util::MakeCallback(completion));
+  _firestore->ClearPersistence(MakeCallback(completion));
 }
 
 - (void)waitForPendingWritesWithCompletion:(void (^)(NSError *_Nullable error))completion {
-  _firestore->WaitForPendingWrites(util::MakeCallback(completion));
+  _firestore->WaitForPendingWrites(MakeCallback(completion));
 }
 
 - (void)terminateWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
   id<FSTFirestoreInstanceRegistry> strongRegistry = _registry;
   if (strongRegistry) {
     [strongRegistry
-        removeInstanceWithDatabase:util::MakeNSString(_firestore->database_id().database_id())];
+        removeInstanceWithDatabase:MakeNSString(_firestore->database_id().database_id())];
   }
   [self terminateInternalWithCompletion:completion];
 }
@@ -367,16 +370,12 @@ NS_ASSUME_NONNULL_BEGIN
   return _firestore->database_id();
 }
 
-+ (BOOL)isLoggingEnabled {
-  return util::LogIsLoggable(util::kLogLevelDebug);
-}
-
 + (FIRFirestore *)recoverFromFirestore:(std::shared_ptr<Firestore>)firestore {
   return (__bridge FIRFirestore *)firestore->extension();
 }
 
 - (void)terminateInternalWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
-  _firestore->Terminate(util::MakeCallback(completion));
+  _firestore->Terminate(MakeCallback(completion));
 }
 
 @end