+++ Firestore/core/src/firebase/firestore/core/database_info.h
@@ -28,12 +28,6 @@ namespace core {
 /** DatabaseInfo contains data about the database. */
 class DatabaseInfo {
  public:
-#if defined(__OBJC__)
-  // For objective-c++ initialization; to be removed after migration.
-  // Do NOT use in C++ code.
-  DatabaseInfo() = default;
-#endif  // defined(__OBJC__)
-
   /**
    * Creates a new DatabaseInfo.
    *
@@ -47,6 +41,8 @@ class DatabaseInfo {
                std::string persistence_key,
                std::string host,
                bool ssl_enabled);
+
+  DatabaseInfo() = default;
 
   const model::DatabaseId& database_id() const {
     return database_id_;
@@ -68,7 +64,7 @@ class DatabaseInfo {
   model::DatabaseId database_id_;
   std::string persistence_key_;
   std::string host_;
-  bool ssl_enabled_;
+  bool ssl_enabled_ = false;
 };
 
 }  // namespace core