+++ Firestore/core/src/firebase/firestore/model/precondition.h
@@ -20,10 +20,6 @@
 #include <string>
 #include <utility>
 
-#if defined(__OBJC__)
-#include "Firestore/core/include/firebase/firestore/timestamp.h"
-#endif  // defined(__OBJC__)
-
 #include "Firestore/core/src/firebase/firestore/model/maybe_document.h"
 #include "Firestore/core/src/firebase/firestore/model/snapshot_version.h"
 #include "Firestore/core/src/firebase/firestore/util/hard_assert.h"