+++ Firestore/core/src/firebase/firestore/remote/grpc_stream.h
@@ -127,7 +127,7 @@ class GrpcStream : public GrpcCall {
              const std::shared_ptr<util::AsyncQueue>& worker_queue,
              GrpcConnection* grpc_connection,
              GrpcStreamObserver* observer);
-  ~GrpcStream();
+  ~GrpcStream() override;
 
   void Start();
 
@@ -196,11 +196,11 @@ class GrpcStream : public GrpcCall {
   void OnRead(const grpc::ByteBuffer& message);
   void OnWrite();
   void OnOperationFailed();
-  void RemoveCompletion(const GrpcCompletion* to_remove);
+  void RemoveCompletion(const std::shared_ptr<GrpcCompletion>& to_remove);
 
-  using OnSuccess = std::function<void(const GrpcCompletion*)>;
-  GrpcCompletion* NewCompletion(GrpcCompletion::Type type,
-                                const OnSuccess& callback);
+  using OnSuccess = std::function<void(const std::shared_ptr<GrpcCompletion>&)>;
+  std::shared_ptr<GrpcCompletion> NewCompletion(GrpcCompletion::Type type,
+                                                const OnSuccess& callback);
   // Finishes the underlying gRPC call. Must always be invoked on any call that
   // was started. Presumes that any pending completions will quickly come off
   // the queue and will block until they do, so this must only be invoked when
@@ -235,7 +235,7 @@ class GrpcStream : public GrpcCall {
   GrpcStreamObserver* observer_ = nullptr;
   internal::BufferedWriter buffered_writer_;
 
-  std::vector<GrpcCompletion*> completions_;
+  std::vector<std::shared_ptr<GrpcCompletion>> completions_;
 
   // gRPC asserts that a call is finished exactly once.
   bool is_grpc_call_finished_ = false;