+++ Firestore/core/src/firebase/firestore/remote/grpc_unary_call.h
@@ -102,7 +102,7 @@ class GrpcUnaryCall : public GrpcCall {
   std::shared_ptr<util::AsyncQueue> worker_queue_;
   GrpcConnection* grpc_connection_ = nullptr;
 
-  GrpcCompletion* finish_completion_ = nullptr;
+  std::shared_ptr<GrpcCompletion> finish_completion_;
   Callback callback_;
 };