+++ Firestore/core/src/firebase/firestore/util/string_apple.h
@@ -48,20 +48,6 @@ inline CFStringRef MakeCFString(absl::string_view contents) {
 
 #if defined(__OBJC__)
 
-// Translates a C string to the equivalent NSString without making a copy.
-inline NSString* MakeNSStringNoCopy(const char* c_str, size_t size) {
-  return [[NSString alloc]
-      initWithBytesNoCopy:const_cast<void*>(static_cast<const void*>(c_str))
-                   length:size
-                 encoding:NSUTF8StringEncoding
-             freeWhenDone:false];
-}
-
-// Translates a string_view to the equivalent NSString without making a copy.
-inline NSString* MakeNSStringNoCopy(absl::string_view str) {
-  return MakeNSStringNoCopy(str.data(), str.size());
-}
-
 // Translates a string_view string to the equivalent NSString by making a copy.
 inline NSString* MakeNSString(absl::string_view str) {
   return [[NSString alloc]